Add filters for reducing amount of data output
This change adds commented-out samples of how to use filters to reduce the amount of data that are written to the HA recorder database by default.

Depending on the `pulse_rate` variable, the type of house/apartment and the heating system in use, this may or may not be a problem in practice. For example, with the default in the file, `1000`, a power consumption of 3600 W means that the sensors produce 2 HA state changes per second (which means 7200 state changes per hour). With the commented-out filters enabled, only 396 state changes will be produced per hour.